refactor: move DatahubSubtype Enum into entities.py
